25/3/2023 Sửa load ảnh, hình nền và floating button ẩn hiện có vấn đề
