+++ finedining/app/src/debug/res/drawable/ic_launcher_foreground.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="108dp"
+    android:height="108dp"
+    android:viewportWidth="108"
+    android:viewportHeight="108">
+  <group android:scaleX="0.23250224"
+      android:scaleY="0.23250224"
+      android:translateX="21.606045"
+      android:translateY="21.6">
+
+      <path
+          android:fillColor="#C0590E"
+          android:fillType="evenOdd"
+          android:pathData="M110.913,275.535C47.784,262.792-4.51,204.195,0.309,130.202 C5.039,57.564,63.771,2.86,133.68,0.111c74.29-2.921,136.473,52.256,144.16,124.282c7.697,72.13-40.244,130.791-96.064,147.288 c-0.475-65.051-0.949-129.97-1.423-194.887c-0.487-0.093-0.974-0.185-1.46-0.277c-0.584,1.033-1.26,2.025-1.74,3.104 c-11.867,26.66-18.539,54.548-19.704,83.725c-0.374,9.391,2.892,17.883,7.402,25.784c2.885,5.052,4.001,10.016,3.763,15.778 c-0.933,22.492-1.588,44.998-2.175,67.505c-0.078,2.985-1.032,4.113-3.937,4.312c-8.25,0.564-16.479,1.438-24.731,1.938 c-3.188,0.193-6.417-0.291-9.798-0.476c-0.143-1.294-0.331-2.29-0.353-3.288c-0.729-33.115-1.428-66.232-2.19-99.346 c-0.048-2.097,0.412-3.234,2.65-3.904c8.589-2.57,13.41-8.617,13.271-17.457c-0.268-17.189-1.127-34.371-1.896-51.547 c-0.277-6.198-0.977-12.378-1.541-18.562c-0.128-1.408-0.374-2.833-0.811-4.171c-0.205-0.627-1.089-1.56-1.495-1.48 c-0.728,0.142-1.776,0.854-1.9,1.485c-0.472,2.389-0.824,4.838-0.871,7.271c-0.295,15.648-0.459,31.297-0.782,46.943 c-0.042,2.054-0.744,4.095-1.141,6.142c-0.467-0.008-0.934-0.017-1.4-0.026c-0.461-2.336-1.243-4.661-1.331-7.011 c-0.575-15.38-0.981-30.767-1.496-46.149c-0.077-2.312-0.32-4.636-0.734-6.909c-0.124-0.679-1.094-1.684-1.664-1.675 c-0.636,0.009-1.715,0.937-1.806,1.578c-0.418,2.926-0.726,5.89-0.767,8.844c-0.208,14.87-0.264,29.741-0.526,44.61 c-0.038,2.161-0.898,4.307-1.378,6.46c-0.453-0.001-0.905-0.001-1.358-0.001c-0.42-1.643-1.176-3.279-1.208-4.93 c-0.293-15.124-0.451-30.251-0.672-45.376c-0.04-2.712,0.023-5.457-0.39-8.12c-0.18-1.161-1.44-2.155-2.209-3.226 c-0.722,1.077-2.016,2.125-2.071,3.233c-0.352,7.101-0.336,14.219-0.622,21.324c-0.44,10.981-0.992,21.958-1.589,32.933 c-0.081,1.489-0.693,2.95-2.016,4.581c-0.455-1.45-1.268-2.891-1.307-4.352c-0.419-15.25-0.699-30.504-1.053-45.755 c-0.072-3.096-0.158-6.204-0.526-9.272c-0.115-0.961-1.177-1.809-1.805-2.708c-0.747,0.915-2.039,1.765-2.155,2.754 c-0.813,6.932-1.57,13.881-1.993,20.846c-0.902,14.847-1.781,29.7-2.249,44.565c-0.141,4.434,0.568,9.077,1.859,13.331 c1.728,5.689,5.738,9.908,11.557,11.7c2.8,0.86,2.695,2.566,2.649,4.695c-0.595,27.425-1.185,54.851-1.728,82.276 C110.819,264.025,110.913,269.564,110.913,275.535z M88.953,235.372c1.057-3.044,0.04-4.667-2.518-6.177 c-13.706-8.094-24.918-18.908-33.558-32.247c-24.391-37.653-22.921-84.031,3.662-120.478 c40.947-56.143,127.21-56.318,168.323-0.343c35.527,48.37,25.142,114.321-23.553,149.102c-2.497,1.784-2.749,3.593-1.729,6.167 c36.314-22.137,61.332-70.671,46.49-122.503c-14.016-48.946-61.408-82.398-111.457-79.039 C81.836,33.396,39.518,72.221,32.2,124.358C25.021,175.5,54.921,218.57,88.953,235.372z" />
+  </group>
+</vector>
