--- finedining/app/src/main/res/drawable/ic_launcher_foreground.xml
@@ -1,16 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="108dp"
-    android:height="108dp"
-    android:viewportWidth="108"
-    android:viewportHeight="108">
-  <group android:scaleX="0.16933328"
-      android:scaleY="0.16933328"
-      android:translateX="-17.28"
-      android:translateY="3.5996408">
-      <path
-          android:pathData="M384.356 463.226c-81.29 -16.409 -148.627 -91.863 -142.422 -187.141c6.091 -93.533 81.719 -163.974 171.739 -167.514c95.66 -3.761 175.732 67.288 185.63 160.035c9.911 92.878 -51.821 168.416 -123.699 189.659c-0.611 -83.765 -1.222 -167.358 -1.832 -250.951c-0.627 -0.119 -1.254 -0.237 -1.881 -0.356c-0.752 1.33 -1.621 2.607 -2.239 3.997c-15.281 34.329 -23.873 70.24 -25.373 107.81c-0.482 12.092 3.723 23.026 9.531 33.201c3.715 6.506 5.152 12.896 4.845 20.316c-1.2 28.964 -2.044 57.944 -2.8 86.925c-0.101 3.844 -1.33 5.297 -5.069 5.553c-10.622 0.727 -21.219 1.851 -31.845 2.493c-4.105 0.249 -8.263 -0.374 -12.617 -0.611c-0.184 -1.666 -0.426 -2.948 -0.454 -4.234c-0.939 -42.642 -1.839 -85.285 -2.82 -127.925c-0.062 -2.699 0.53 -4.164 3.412 -5.026c11.06 -3.31 17.268 -11.097 17.09 -22.479c-0.345 -22.134 -1.451 -44.258 -2.442 -66.376c-0.357 -7.981 -1.258 -15.939 -1.984 -23.901c-0.166 -1.813 -0.481 -3.648 -1.044 -5.371c-0.264 -0.808 -1.402 -2.008 -1.924 -1.906c-0.937 0.183 -2.287 1.101 -2.447 1.913c-0.608 3.077 -1.061 6.23 -1.121 9.362c-0.381 20.149 -0.591 40.301 -1.007 60.447c-0.054 2.645 -0.958 5.273 -1.469 7.909c-0.601 -0.011 -1.202 -0.022 -1.803 -0.034c-0.593 -3.009 -1.6 -6.001 -1.713 -9.028c-0.741 -19.804 -1.263 -39.617 -1.926 -59.424c-0.099 -2.977 -0.413 -5.969 -0.945 -8.896c-0.159 -0.875 -1.409 -2.167 -2.143 -2.157c-0.818 0.012 -2.208 1.206 -2.326 2.032c-0.538 3.768 -0.935 7.584 -0.987 11.388c-0.267 19.147 -0.34 38.297 -0.677 57.442c-0.049 2.783 -1.157 5.546 -1.774 8.318c-0.583 -0.001 -1.166 -0.001 -1.749 -0.002c-0.542 -2.115 -1.515 -4.223 -1.557 -6.348c-0.377 -19.475 -0.581 -38.953 -0.866 -58.43c-0.051 -3.493 0.03 -7.026 -0.501 -10.456c-0.232 -1.495 -1.854 -2.776 -2.845 -4.154c-0.929 1.387 -2.596 2.736 -2.667 4.164c-0.453 9.143 -0.433 18.309 -0.8 27.458c-0.567 14.14 -1.278 28.274 -2.046 42.406c-0.104 1.917 -0.893 3.799 -2.596 5.898c-0.585 -1.867 -1.632 -3.723 -1.683 -5.604c-0.539 -19.638 -0.9 -39.279 -1.355 -58.918c-0.093 -3.986 -0.204 -7.988 -0.678 -11.939c-0.148 -1.237 -1.515 -2.329 -2.324 -3.488c-0.962 1.178 -2.626 2.273 -2.775 3.546c-1.047 8.926 -2.021 17.874 -2.566 26.842c-1.161 19.118 -2.292 38.244 -2.896 57.385c-0.181 5.71 0.731 11.688 2.394 17.165c2.224 7.326 7.389 12.76 14.882 15.066c3.605 1.108 3.47 3.305 3.411 6.046c-0.766 35.314 -1.526 70.63 -2.224 105.945C384.235 448.404 384.356 455.537 384.356 463.226zM356.078 411.509c1.361 -3.92 0.052 -6.01 -3.242 -7.954c-17.648 -10.422 -32.086 -24.348 -43.212 -41.523c-31.407 -48.485 -29.514 -108.204 4.715 -155.136c52.726 -72.293 163.805 -72.52 216.743 -0.442c45.748 62.285 32.375 147.208 -30.327 191.995c-3.216 2.297 -3.54 4.626 -2.226 7.941c46.761 -28.505 78.975 -91.001 59.863 -157.744c-18.048 -63.026 -79.073 -106.102 -143.519 -101.776C346.915 151.431 292.423 201.425 283 268.56C273.756 334.413 312.257 389.873 356.078 411.509z"
-          android:fillType="evenOdd"
-          android:fillColor="#C0590E" />
-  </group>
-</vector>
