+++ finedining/app/src/main/res/layout/fragment_them_nhan_vien.xml
@@ -16,7 +16,9 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:orientation="vertical"
-                    android:gravity="center">
+                    android:gravity="center"
+                    android:clickable="true"
+                    android:focusable="true">
 
                     <TextView
                         android:id="@+id/tv_fThemNhanVien_title"
